+++ NEWS.md
@@ -1,9 +1,8 @@
 # MSnbase 2.17
 
-## Changes in 2.17.4
-- Add `normalize,Chromatogram` method.
-
 ## Changes in 2.17.3
+- Add `normalize,Chromatogram` method.
+- Add `filterIntensity,Chromatogram` method.
 - Use two digits when printing retention time seconds
 
 ## Changes in 2.17.2